api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
redis[asyncio]==5.0.4
python-dotenv==1.0.1
